refactor(spells): decompose validator.ts along pre-marked seams
validator.ts was 664 lines, over the 500-line ceiling mandated by
CLAUDE.md and flagged during the #517 post-merge review. The file
already had clean internal boundaries (section-comment banners), so
this is mechanical extraction — no behavior changes.

Split:
- validator.ts (133 lines) — thin composition entry, public API only
- validators/top-level.ts — validateTopLevel + validateArguments + matchesArgumentType
- validators/steps.ts — validateSteps + nested preflight validation
- validators/prerequisites.ts — validatePrerequisites (#460)
- validators/references.ts — validateVariableReferences + parallel cross-ref (#247)
- validators/jumps.ts — detectCircularJumps

Also tightened a few things the split made obvious:
- validateVariableReferences.errors is now required; dropped the dead
  `if (!errors) return;` silent-failure guard since the sole caller
  always passes an array.
- steps.ts imports `ValidatorOptions` from the entry instead of
  re-declaring a structurally identical `StepsValidatorOptions`.
- Preflight nested validation extracted to validatePreflight /
  validatePreflightResolutions helpers (shallower nesting, same logic).

Public API surface (`validateSpellDefinition`, `resolveArguments`,
`ValidatorOptions`) is byte-identical — all 11 internal consumers and
the package index.ts re-export continue to work unchanged.

Verified: @moflo/spells 1143/1143 + @moflo/cli 2011/2011 + isolation
batch 481/481, all green; tsc --noEmit clean across the monorepo.

Closes #521
