+++ src/modules/spells/src/schema/validator.ts
@@ -8,24 +8,25 @@
  * - No circular variable references
  * - Valid argument definitions
  * - Undefined step output references detected
+ *
+ * This module is a thin composition entry; the actual rules live under
+ * `./validators/` and each file owns a single concern (top-level,
+ * steps, prerequisites, references, jumps).
  */
 
 import type { ValidationResult, ValidationError } from '../types/step-command.types.js';
 import type {
   SpellDefinition,
-  StepDefinition,
   ArgumentDefinition,
-  ArgumentType,
-  PrerequisiteSpec,
 } from '../types/spell-definition.types.js';
-import { validateStepCapabilities, isValidMofloLevel, compareMofloLevels } from '../core/capability-validator.js';
+import { isValidMofloLevel } from '../core/capability-validator.js';
 import { MOFLO_LEVEL_ORDER } from '../types/step-command.types.js';
-import type { MofloLevel } from '../types/step-command.types.js';
-import { isValidPermissionLevel, VALID_PERMISSION_LEVELS } from '../core/permission-resolver.js';
 import { validateSchedule } from '../scheduler/cron-parser.js';
-import { VAR_REF_PATTERN } from '../core/interpolation.js';
-
-const VALID_ARG_TYPES: readonly ArgumentType[] = ['string', 'number', 'boolean', 'string[]'];
+import { validateTopLevel, validateArguments, matchesArgumentType } from './validators/top-level.js';
+import { validateSteps } from './validators/steps.js';
+import { validatePrerequisites } from './validators/prerequisites.js';
+import { validateVariableReferences } from './validators/references.js';
+import { detectCircularJumps } from './validators/jumps.js';
 
 export interface ValidatorOptions {
   /** Known step command types. If provided, unknown types produce an error. */
@@ -43,7 +44,6 @@ export function validateSpellDefinition(
 
   validateTopLevel(def, errors);
 
-  // Validate spell-level mofloLevel
   if (def.mofloLevel !== undefined && !isValidMofloLevel(def.mofloLevel)) {
     errors.push({
       path: 'mofloLevel',
@@ -70,447 +70,6 @@ export function validateSpellDefinition(
 
   return { valid: errors.length === 0, errors };
 }
-
-// ============================================================================
-// Top-level validation
-// ============================================================================
-
-function validateTopLevel(def: SpellDefinition, errors: ValidationError[]): void {
-  if (!def.name || typeof def.name !== 'string') {
-    errors.push({ path: 'name', message: 'name is required and must be a string' });
-  }
-  if (!def.steps || !Array.isArray(def.steps) || def.steps.length === 0) {
-    errors.push({ path: 'steps', message: 'steps is required and must be a non-empty array' });
-  }
-}
-
-// ============================================================================
-// Argument validation
-// ============================================================================
-
-function validateArguments(
-  args: Record<string, ArgumentDefinition>,
-  errors: ValidationError[],
-): void {
-  for (const [name, argDef] of Object.entries(args)) {
-    const path = `arguments.${name}`;
-
-    if (!argDef.type || !VALID_ARG_TYPES.includes(argDef.type)) {
-      errors.push({
-        path: `${path}.type`,
-        message: `type must be one of: ${VALID_ARG_TYPES.join(', ')}`,
-      });
-    }
-
-    if (argDef.default !== undefined && argDef.type) {
-      if (!matchesArgumentType(argDef.default, argDef.type)) {
-        errors.push({
-          path: `${path}.default`,
-          message: `default value ${JSON.stringify(argDef.default)} does not match declared type "${argDef.type}"`,
-        });
-      }
-    }
-
-    if (argDef.enum !== undefined) {
-      if (!Array.isArray(argDef.enum) || argDef.enum.length === 0) {
-        errors.push({ path: `${path}.enum`, message: 'enum must be a non-empty array' });
-      } else if (argDef.type) {
-        for (const enumVal of argDef.enum) {
-          if (!matchesArgumentType(enumVal, argDef.type)) {
-            errors.push({
-              path: `${path}.enum`,
-              message: `enum value ${JSON.stringify(enumVal)} does not match declared type "${argDef.type}"`,
-            });
-            break; // one error per enum is sufficient
-          }
-        }
-      }
-    }
-
-    if (argDef.default !== undefined && argDef.enum !== undefined) {
-      if (!argDef.enum.includes(argDef.default)) {
-        errors.push({
-          path: `${path}.default`,
-          message: `default value "${argDef.default}" is not in enum: ${argDef.enum.join(', ')}`,
-        });
-      }
-    }
-  }
-}
-
-// ============================================================================
-// Step validation
-// ============================================================================
-
-function validateSteps(
-  steps: readonly StepDefinition[],
-  errors: ValidationError[],
-  stepIds: Set<string>,
-  outputVars: Set<string>,
-  options?: ValidatorOptions,
-  prefix = 'steps',
-  spellLevel?: MofloLevel,
-): void {
-  for (let i = 0; i < steps.length; i++) {
-    const step = steps[i];
-    const path = `${prefix}[${i}]`;
-
-    if (!step.id || typeof step.id !== 'string') {
-      errors.push({ path: `${path}.id`, message: 'step id is required' });
-    } else if (stepIds.has(step.id)) {
-      errors.push({ path: `${path}.id`, message: `duplicate step id: "${step.id}"` });
-    } else {
-      stepIds.add(step.id);
-    }
-
-    if (!step.type || typeof step.type !== 'string') {
-      errors.push({ path: `${path}.type`, message: 'step type is required' });
-    } else if (options?.knownStepTypes && !options.knownStepTypes.includes(step.type)) {
-      const suggestions = findSimilar(step.type, options.knownStepTypes);
-      const hint = suggestions.length > 0 ? ` Did you mean: ${suggestions.join(', ')}?` : '';
-      errors.push({
-        path: `${path}.type`,
-        message: `unknown step type: "${step.type}".${hint}`,
-      });
-    }
-
-    if (step.output) {
-      outputVars.add(step.output);
-    }
-
-    // Validate capabilities if declared
-    const capErrors = validateStepCapabilities(step, path);
-    errors.push(...capErrors);
-
-    // Validate mofloLevel if declared
-    if (step.mofloLevel !== undefined) {
-      if (!isValidMofloLevel(step.mofloLevel)) {
-        errors.push({
-          path: `${path}.mofloLevel`,
-          message: `invalid mofloLevel: "${step.mofloLevel}". Valid levels: ${MOFLO_LEVEL_ORDER.join(', ')}`,
-        });
-      } else if (spellLevel && compareMofloLevels(step.mofloLevel, spellLevel) > 0) {
-        errors.push({
-          path: `${path}.mofloLevel`,
-          message: `step mofloLevel "${step.mofloLevel}" exceeds spell-level "${spellLevel}" — steps can only narrow, not escalate`,
-        });
-      }
-    }
-
-    // Validate permissionLevel if declared
-    if (step.permissionLevel !== undefined && !isValidPermissionLevel(step.permissionLevel)) {
-      errors.push({
-        path: `${path}.permissionLevel`,
-        message: `invalid permissionLevel: "${step.permissionLevel}". Valid levels: ${VALID_PERMISSION_LEVELS.join(', ')}`,
-      });
-    }
-
-    // Validate declarative prerequisite specs (external deps)
-    if (step.prerequisites !== undefined) {
-      validatePrerequisites(step.prerequisites, errors, `${path}.prerequisites`);
-    }
-
-    // Validate declarative preflight specs (runtime state checks)
-    if (step.preflight !== undefined) {
-      if (!Array.isArray(step.preflight)) {
-        errors.push({ path: `${path}.preflight`, message: 'preflight must be an array' });
-      } else {
-        step.preflight.forEach((pf, pi) => {
-          const pfPath = `${path}.preflight[${pi}]`;
-          if (!pf || typeof pf !== 'object') {
-            errors.push({ path: pfPath, message: 'preflight entry must be an object' });
-            return;
-          }
-          if (typeof pf.name !== 'string' || pf.name.length === 0) {
-            errors.push({ path: `${pfPath}.name`, message: 'preflight.name is required' });
-          }
-          if (typeof pf.command !== 'string' || pf.command.length === 0) {
-            errors.push({ path: `${pfPath}.command`, message: 'preflight.command is required' });
-          }
-          if (pf.expectExitCode !== undefined && typeof pf.expectExitCode !== 'number') {
-            errors.push({ path: `${pfPath}.expectExitCode`, message: 'expectExitCode must be a number' });
-          }
-          if (pf.timeoutMs !== undefined && (typeof pf.timeoutMs !== 'number' || pf.timeoutMs <= 0)) {
-            errors.push({ path: `${pfPath}.timeoutMs`, message: 'timeoutMs must be a positive number' });
-          }
-          if (pf.hint !== undefined && typeof pf.hint !== 'string') {
-            errors.push({ path: `${pfPath}.hint`, message: 'hint must be a string' });
-          }
-          if (pf.severity !== undefined && pf.severity !== 'fatal' && pf.severity !== 'warning') {
-            errors.push({ path: `${pfPath}.severity`, message: 'severity must be "fatal" or "warning"' });
-          }
-          if (pf.resolutions !== undefined) {
-            if (!Array.isArray(pf.resolutions)) {
-              errors.push({ path: `${pfPath}.resolutions`, message: 'resolutions must be an array' });
-            } else {
-              pf.resolutions.forEach((r: Record<string, unknown>, ri: number) => {
-                const rPath = `${pfPath}.resolutions[${ri}]`;
-                if (!r || typeof r !== 'object') {
-                  errors.push({ path: rPath, message: 'resolution must be an object' });
-                  return;
-                }
-                if (typeof r.label !== 'string' || r.label.length === 0) {
-                  errors.push({ path: `${rPath}.label`, message: 'resolution.label is required' });
-                }
-                if (r.command !== undefined && typeof r.command !== 'string') {
-                  errors.push({ path: `${rPath}.command`, message: 'resolution.command must be a string' });
-                }
-                if (r.timeoutMs !== undefined && (typeof r.timeoutMs !== 'number' || r.timeoutMs <= 0)) {
-                  errors.push({ path: `${rPath}.timeoutMs`, message: 'resolution.timeoutMs must be a positive number' });
-                }
-              });
-            }
-          }
-        });
-      }
-    }
-
-    // Recurse into nested steps (condition/loop/parallel)
-    if (step.steps && Array.isArray(step.steps)) {
-      validateSteps(step.steps, errors, stepIds, outputVars, options, `${path}.steps`, spellLevel);
-
-      // Parallel blocks: reject cross-references between sibling steps (#247)
-      if (step.type === 'parallel') {
-        validateParallelCrossReferences(step.steps, errors, `${path}.steps`);
-      }
-    }
-  }
-}
-
-// ============================================================================
-// Prerequisite spec validation
-// ============================================================================
-
-const VALID_DETECT_TYPES = ['env', 'command', 'file'] as const;
-
-function validatePrerequisites(
-  prereqs: readonly PrerequisiteSpec[],
-  errors: ValidationError[],
-  path: string,
-): void {
-  if (!Array.isArray(prereqs)) {
-    errors.push({ path, message: 'prerequisites must be an array' });
-    return;
-  }
-
-  const seenNames = new Set<string>();
-  prereqs.forEach((p, i) => {
-    const pPath = `${path}[${i}]`;
-    if (!p || typeof p !== 'object') {
-      errors.push({ path: pPath, message: 'prerequisite entry must be an object' });
-      return;
-    }
-    if (typeof p.name !== 'string' || p.name.length === 0) {
-      errors.push({ path: `${pPath}.name`, message: 'prerequisite.name is required' });
-    } else if (seenNames.has(p.name)) {
-      errors.push({
-        path: `${pPath}.name`,
-        message: `duplicate prerequisite name "${p.name}" in the same block`,
-      });
-    } else {
-      seenNames.add(p.name);
-    }
-    if (p.description !== undefined && typeof p.description !== 'string') {
-      errors.push({ path: `${pPath}.description`, message: 'description must be a string' });
-    }
-    if (p.docsUrl !== undefined && typeof p.docsUrl !== 'string') {
-      errors.push({ path: `${pPath}.docsUrl`, message: 'docsUrl must be a string' });
-    }
-    if (p.promptOnMissing !== undefined && typeof p.promptOnMissing !== 'boolean') {
-      errors.push({ path: `${pPath}.promptOnMissing`, message: 'promptOnMissing must be a boolean' });
-    }
-
-    const detect = p.detect as PrerequisiteSpec['detect'] | undefined;
-    if (!detect || typeof detect !== 'object') {
-      errors.push({ path: `${pPath}.detect`, message: 'detect is required and must be an object' });
-      return;
-    }
-    if (!VALID_DETECT_TYPES.includes(detect.type as typeof VALID_DETECT_TYPES[number])) {
-      errors.push({
-        path: `${pPath}.detect.type`,
-        message: `detect.type must be one of: ${VALID_DETECT_TYPES.join(', ')}`,
-      });
-      return;
-    }
-    if (detect.type === 'env') {
-      if (typeof detect.key !== 'string' || detect.key.length === 0) {
-        errors.push({ path: `${pPath}.detect.key`, message: 'detect.key is required for env detector' });
-      }
-    } else if (detect.type === 'command') {
-      if (typeof detect.command !== 'string' || detect.command.length === 0) {
-        errors.push({ path: `${pPath}.detect.command`, message: 'detect.command is required for command detector' });
-      }
-    } else if (detect.type === 'file') {
-      if (typeof detect.path !== 'string' || detect.path.length === 0) {
-        errors.push({ path: `${pPath}.detect.path`, message: 'detect.path is required for file detector' });
-      }
-    }
-  });
-}
-
-// ============================================================================
-// Variable reference validation
-// ============================================================================
-
-/**
- * Check that {stepId.outputKey} references point to declared step outputs.
- * Also detects forward references (referencing a step that hasn't executed yet).
- */
-function validateVariableReferences(
-  steps: readonly StepDefinition[],
-  outputVars: Set<string>,
-  args?: Record<string, ArgumentDefinition>,
-  errors?: ValidationError[],
-  prefix = 'steps',
-  parentDeclaredStepIds: readonly string[] = [],
-): void {
-  if (!errors) return;
-  // Nested steps can reference parent-scope declarations (e.g. a loop body
-  // referencing an output declared before the loop).
-  const declaredStepIds: string[] = [...parentDeclaredStepIds];
-
-  for (let i = 0; i < steps.length; i++) {
-    const step = steps[i];
-    const path = `${prefix}[${i}]`;
-
-    // Check string config values for variable references
-    if (step.config) {
-      for (const [key, value] of Object.entries(step.config)) {
-        if (typeof value === 'string') {
-          checkStringReferences(value, declaredStepIds, args, errors, `${path}.config.${key}`);
-        }
-      }
-    }
-
-    // After checking, this step's ID and output variable are available to subsequent steps
-    if (step.id) {
-      declaredStepIds.push(step.id);
-    }
-    if (step.output && typeof step.output === 'string' && step.output !== step.id) {
-      declaredStepIds.push(step.output);
-    }
-
-    // Recurse into nested steps (condition/loop bodies)
-    if (step.steps && Array.isArray(step.steps)) {
-      validateVariableReferences(step.steps, outputVars, args, errors, `${path}.steps`, declaredStepIds);
-
-      // Parallel/loop nested step IDs are available to subsequent top-level steps (#247)
-      for (const nested of step.steps) {
-        if (nested.id) declaredStepIds.push(nested.id);
-      }
-    }
-  }
-}
-
-
-function checkStringReferences(
-  value: string,
-  declaredStepIds: string[],
-  args?: Record<string, ArgumentDefinition>,
-  errors?: ValidationError[],
-  path?: string,
-): void {
-  if (!errors || !path) return;
-
-  VAR_REF_PATTERN.lastIndex = 0;
-  let match;
-  while ((match = VAR_REF_PATTERN.exec(value)) !== null) {
-    const ref = match[1];
-    const segments = ref.split('.');
-
-    // {args.key} — check arg is defined
-    if (segments[0] === 'args' && segments.length === 2) {
-      if (args && !args[segments[1]]) {
-        errors.push({
-          path,
-          message: `references undefined argument: "${segments[1]}"`,
-        });
-      }
-      continue;
-    }
-
-    // {credentials.NAME} — resolved at runtime by the runner
-    if (segments[0] === 'credentials' && segments.length === 2) {
-      continue;
-    }
-
-    // {env.KEY} — resolved at runtime from process.env, typically populated
-    // by a spell-level `prerequisites:` entry with promptOnMissing: true.
-    if (segments[0] === 'env' && segments.length === 2) {
-      continue;
-    }
-
-    // {loop.varName} — resolved at runtime by the loop executor
-    if (segments[0] === 'loop') {
-      continue;
-    }
-
-    // {stepId.outputKey} — check step has been declared before this point
-    if (segments.length >= 2) {
-      const stepId = segments[0];
-      if (!declaredStepIds.includes(stepId)) {
-        // Check if it's a forward reference to a later step
-        errors.push({
-          path,
-          message: `references step "${stepId}" which has not been declared yet (forward reference)`,
-        });
-      }
-    }
-  }
-}
-
-// ============================================================================
-// Parallel cross-reference validation (#247)
-// ============================================================================
-
-/**
- * Reject variable references between sibling steps within a parallel block.
- * Steps in a parallel block execute concurrently, so they cannot depend on
- * each other's outputs (execution order is undefined).
- */
-function validateParallelCrossReferences(
-  steps: readonly StepDefinition[],
-  errors: ValidationError[],
-  prefix: string,
-): void {
-  const siblingIds = new Set<string>();
-  for (const step of steps) {
-    if (step.id) siblingIds.add(step.id);
-  }
-
-  for (let i = 0; i < steps.length; i++) {
-    const step = steps[i];
-    const path = `${prefix}[${i}]`;
-    if (!step.config) continue;
-
-    for (const [key, value] of Object.entries(step.config)) {
-      if (typeof value !== 'string') continue;
-      VAR_REF_PATTERN.lastIndex = 0;
-      let match;
-      while ((match = VAR_REF_PATTERN.exec(value)) !== null) {
-        const ref = match[1];
-        const segments = ref.split('.');
-        if (
-          segments.length >= 2
-          && segments[0] !== 'args'
-          && segments[0] !== 'credentials'
-          && segments[0] !== 'env'
-        ) {
-          if (siblingIds.has(segments[0]) && segments[0] !== step.id) {
-            errors.push({
-              path: `${path}.config.${key}`,
-              message: `parallel step "${step.id}" references sibling step "${segments[0]}" — steps within a parallel block cannot reference each other's outputs`,
-            });
-          }
-        }
-      }
-    }
-  }
-}
-
-// ============================================================================
-// Argument resolution
-// ============================================================================
 
 /**
  * Resolve provided arguments against definitions, applying defaults and validation.
@@ -571,105 +130,4 @@ export function resolveArguments(
   }
 
   return { resolved, errors };
-}
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-/** Check whether a value matches a declared ArgumentType. */
-function matchesArgumentType(value: unknown, type: ArgumentType): boolean {
-  switch (type) {
-    case 'string': return typeof value === 'string';
-    case 'number': return typeof value === 'number';
-    case 'boolean': return typeof value === 'boolean';
-    case 'string[]': return Array.isArray(value) && value.every(v => typeof v === 'string');
-    default: return true;
-  }
-}
-
-function findSimilar(target: string, candidates: readonly string[]): string[] {
-  return candidates
-    .filter(c => {
-      // Simple check: shares at least 3 characters or starts with same prefix
-      const common = [...target].filter(ch => c.includes(ch)).length;
-      return common >= 3 || c.startsWith(target.slice(0, 3));
-    })
-    .slice(0, 3);
-}
-
-// ============================================================================
-// Circular step jump detection
-// ============================================================================
-
-/**
- * Detect cycles in condition step jump targets (then/else).
- * Uses DFS with a visiting set to find back-edges.
- */
-function detectCircularJumps(
-  steps: readonly StepDefinition[],
-  errors: ValidationError[],
-): void {
-  // Build adjacency: stepId → set of jump target stepIds
-  const edges = new Map<string, Set<string>>();
-  const stepIdSet = new Set<string>();
-
-  for (const step of steps) {
-    if (!step.id) continue;
-    stepIdSet.add(step.id);
-
-    if (step.type === 'condition' && step.config) {
-      const targets = new Set<string>();
-      const thenTarget = step.config.then;
-      const elseTarget = step.config.else;
-      if (typeof thenTarget === 'string' && thenTarget.length > 0) {
-        targets.add(thenTarget);
-      }
-      if (typeof elseTarget === 'string' && elseTarget.length > 0) {
-        targets.add(elseTarget);
-      }
-      if (targets.size > 0) {
-        edges.set(step.id, targets);
-      }
-    }
-  }
-
-  if (edges.size === 0) return;
-
-  // DFS cycle detection
-  const visited = new Set<string>();
-  const visiting = new Set<string>();
-
-  function visit(id: string, path: string[]): void {
-    if (visited.has(id)) return;
-    if (visiting.has(id)) {
-      const cycleStart = path.indexOf(id);
-      const cycle = path.slice(cycleStart).concat(id);
-      errors.push({
-        path: 'steps',
-        message: `circular condition jump detected: ${cycle.join(' → ')}`,
-      });
-      return;
-    }
-
-    visiting.add(id);
-    path.push(id);
-
-    const targets = edges.get(id);
-    if (targets) {
-      for (const target of targets) {
-        if (stepIdSet.has(target)) {
-          visit(target, path);
-        }
-      }
-    }
-
-    path.pop();
-    visiting.delete(id);
-    visited.add(id);
-  }
-
-  for (const id of edges.keys()) {
-    visit(id, []);
-  }
 }
